Added some UT

+++ src/types.rs
@@ -8,7 +8,7 @@ pub type Score = i32;
 #[derive(Clone, Parser)]
 pub struct Parameters {
     #[arg(short = 'i', long = "iterations", default_value = "200")]
-    pub(crate) iterations: i32,
+    pub(crate) iterations: i64,
 
     // #[arg(short = 'v', long = "verbose")]
     // pub(crate) verbose: bool,
